+++ src/ServerContext.h
@@ -3,8 +3,10 @@
 #include <map>
 #include <memory>
 
+#include "open62541pp/services/NodeManagement.h"
 #include "open62541pp/services/Subscription.h"
 #include "open62541pp/types/Composed.h"
+#include "open62541pp/types/NodeId.h"
 
 namespace opcua {
 
@@ -19,7 +21,12 @@ public:
         services::DataChangeNotificationCallback dataChangeCallback;
     };
 
+    struct NodeContext {
+        services::MethodCallback methodCallback;
+    };
+
     std::map<uint32_t, std::unique_ptr<MonitoredItem>> monitoredItems;
+    std::map<NodeId, std::unique_ptr<NodeContext>> nodeContexts;
 };
 
 }  // namespace opcua